fix: ProblemService 페이지 기능 추가
1.  쿼리스트링으로 받은 페이지에 해당하는 문제만 반환하도록 변경

Related to : #9

+++ backend/src/main/java/com/prograngers/backend/controller/ProblemController.java
@@ -23,7 +23,7 @@ public class ProblemController {
     private final ProblemService problemService;
 
     @GetMapping
-    public ResponseEntity<?> problems(@RequestParam Long page) {
+    public ResponseEntity<?> problems(@RequestParam Integer page) {
         List<ProblemListResponse> problemListResponses = problemService.getProblemList(page);
         return ResponseEntity.ok(problemListResponses);
     }
